Replace the config file reading to a separate module.
